+++ src/transactions/Positions/position.test.js
@@ -1,12 +1,30 @@
 import Decimal from 'decimal.js'
 import Position from './position.js'
 
+const posParams = {
+  assetManagerId: 8,
+  bookId: 'book-1',
+  assetId: 'abc',
+  quantity: 1000,
+  averagePrice: 500.04,
+  validFrom: '2017-01-01',
+  internalId: 1,
+  validTo: '2017-07-04',
+  clientId: 1,
+  accountingType: 'Transaction Date',
+  accountId: 'Asset'
+}
+
 describe('Position class', () => {
   describe('serialization', () => {
-    it('should serialize properly', () => {
-      const test = new Position({ quantity: 12, bookId: 'test' })
-      expect(JSON.parse(JSON.stringify(test)).bookId).toBeDefined()
-      expect(JSON.parse(JSON.stringify(test)).bookId).toEqual('test')
+    it('should construct properly', () => {
+      const test = new Position(posParams)
+      const expectedAttrs = {
+        ...posParams,
+        quantity: new Decimal(1000),
+        averagePrice: new Decimal(500.04)
+      }
+      expect({ ...test }).toEqual(expect.objectContaining(expectedAttrs))
     })
   })
   describe('constructor', () => {
@@ -14,6 +32,10 @@ describe('Position class', () => {
       const testPos = new Position({ quantity: 5.66753 })
       expect(testPos.quantity).toBeInstanceOf(Decimal)
       expect(testPos.quantity).toEqual(new Decimal(5.66753))
+    })
+    it('defaults averagePrice to 0', () => {
+      const pos = new Position({ averagePrice: null })
+      expect(pos.averagePrice).toBeInstanceOf(Decimal)
     })
   })
 })